+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+* `adjuster.string().case()` - for case conversion
+
+### Others
+
+* update packages
+
 ## [1.3.0] - 2019-01-26
 
 ### Others