+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 * `adjuster.number().map()`
+* `adjuster.string().map()`
 
 ## [1.3.0] - 2019-01-26