+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * `adjuster.number().map()`
 * `adjuster.string().map()`
+* `adjuster.numericString().map()`
 
 ## [1.3.0] - 2019-01-26